add order fragment function check box ....

+++ app/src/main/java/com/example/shopingapp/model/OrderRequest.kt
@@ -2,5 +2,5 @@ package com.example.shopingapp.model
 data class OrderRequest(
     val userId: Int,
     val address: String,
-    val totalAmount: Int
+    val totalAmount: Double
 )
